refactor: share marketplace install logic through core
