+++ apps/cli/src/commands/mcp.ts
@@ -1,5 +1,11 @@
+import {
+	installMcpServer,
+	type McpInstallResult,
+	type McpServerTransportConfig,
+} from "@cline/core";
 import type { McpAddDefaults } from "../wizards/mcp";
-import { addServer, type McpTransport } from "../wizards/mcp/settings";
+
+export { buildMcpInstallTransport } from "@cline/core";
 
 export interface McpCommandIo {
 	writeln?: (text: string) => void;
@@ -21,13 +27,13 @@ export interface McpInstallOptions {
 export interface McpInstallDirectResult {
 	name: string;
 	status: "installed";
-	transport: McpTransport;
+	transport: McpServerTransportConfig;
 	warnings: string[];
 }
 
 function normalizeTransportType(
 	value: string | undefined,
-): McpTransport["type"] {
+): McpServerTransportConfig["type"] {
 	const normalized = (value ?? "stdio").trim();
 	if (normalized === "http" || normalized === "streamable-http") {
 		return "streamableHttp";
@@ -56,72 +62,6 @@ function assertValidUrl(url: string): void {
 			`Invalid MCP server URL: ${url} (only http and https are supported)`,
 		);
 	}
-}
-
-function parseHeader(value: string): [string, string] {
-	const separatorIndex = value.indexOf(":");
-	if (separatorIndex <= 0) {
-		throw new Error(
-			`Invalid MCP header "${value}". Expected "Header-Name: header value".`,
-		);
-	}
-	const name = value.slice(0, separatorIndex).trim();
-	const headerValue = value.slice(separatorIndex + 1).trim();
-	if (!name || !headerValue) {
-		throw new Error(
-			`Invalid MCP header "${value}". Expected "Header-Name: header value".`,
-		);
-	}
-	if (!/^[!#$%&'*+\-.^_`|~0-9A-Za-z]+$/.test(name)) {
-		throw new Error(`Invalid MCP header name "${name}".`);
-	}
-	return [name, headerValue];
-}
-
-function splitTargetArgsAndHeaders(input: {
-	headers?: string[];
-	targetArgs?: string[];
-}): { headers: string[]; targetArgs: string[] } {
-	const headers = [...(input.headers ?? [])];
-	const targetArgs: string[] = [];
-	const args = input.targetArgs ?? [];
-	for (let index = 0; index < args.length; index++) {
-		const arg = args[index];
-		if (arg === "--header") {
-			const value = args[index + 1];
-			if (!value) {
-				throw new Error("--header requires a value");
-			}
-			headers.push(value);
-			index++;
-			continue;
-		}
-		if (arg?.startsWith("--header=")) {
-			headers.push(arg.slice("--header=".length));
-			continue;
-		}
-		targetArgs.push(arg);
-	}
-	return { headers, targetArgs };
-}
-
-function buildHeaders(values: string[]): {
-	headers?: Record<string, string>;
-	warnings: string[];
-} {
-	if (values.length === 0) return { warnings: [] };
-	const headers: Record<string, string> = {};
-	const warnings: string[] = [];
-	for (const value of values) {
-		const [name, headerValue] = parseHeader(value);
-		headers[name] = headerValue;
-		if (/<[^>]+>/.test(headerValue)) {
-			warnings.push(
-				`Header "${name}" looks like it contains a placeholder. Update it in MCP settings before using this server.`,
-			);
-		}
-	}
-	return { headers, warnings };
 }
 
 function quoteCommandArg(arg: string): string {
@@ -169,67 +109,15 @@ export function buildMcpInstallDefaults(options: {
 	};
 }
 
-export function buildMcpInstallTransport(options: {
-	headers?: string[];
-	name: string;
-	targetArgs?: string[];
-	transport?: string;
-}): { name: string; transport: McpTransport; warnings: string[] } {
-	const name = options.name.trim();
-	if (!name) {
-		throw new Error("MCP server name is required");
-	}
-	const type = normalizeTransportType(options.transport);
-	const { headers: rawHeaders, targetArgs } = splitTargetArgsAndHeaders({
-		headers: options.headers,
-		targetArgs: options.targetArgs,
-	});
-	const { headers, warnings } = buildHeaders(rawHeaders);
-	if (type === "stdio") {
-		if (rawHeaders.length > 0) {
-			throw new Error("Stdio MCP installs do not support request headers.");
-		}
-		const [command, ...args] = targetArgs;
-		if (!command?.trim()) {
-			throw new Error(
-				"Stdio MCP install requires a command after the server name, for example: cline mcp install fs --yes -- npx -y @modelcontextprotocol/server-filesystem /tmp",
-			);
-		}
-		return {
-			name,
-			transport: {
-				type,
-				command,
-				args: args.length > 0 ? args : undefined,
-			},
-			warnings,
-		};
-	}
-
-	if (targetArgs.length !== 1) {
-		throw new Error(
-			"Remote MCP install requires exactly one URL argument after the server name.",
-		);
-	}
-	const url = targetArgs[0]?.trim() ?? "";
-	assertValidUrl(url);
-	return {
-		name,
-		transport: headers ? { type, url, headers } : { type, url },
-		warnings,
-	};
-}
-
 export function installMcpServerDirect(
 	options: McpInstallOptions,
 ): McpInstallDirectResult {
-	const { name, transport, warnings } = buildMcpInstallTransport(options);
-	addServer(name, transport);
+	const result: McpInstallResult = installMcpServer(options);
 	return {
-		name,
-		status: "installed",
-		transport,
-		warnings,
+		name: result.name,
+		status: result.status,
+		transport: result.transport,
+		warnings: result.warnings,
 	};
 }
