+++ apps/vscode/src/core/controller/marketplace/installMarketplaceEntry.ts
@@ -1,6 +1,6 @@
 import { type MarketplaceEntryRequest, MarketplaceInstallResult } from "@shared/proto/cline/marketplace"
 import type { Controller } from "../index"
-import { installMarketplaceEntryWithCli } from "./marketplace-helpers"
+import { installMarketplaceEntryFromCatalog } from "./marketplace-helpers"
 
 export async function installMarketplaceEntry(
 	_controller: Controller,
@@ -9,5 +9,5 @@ export async function installMarketplaceEntry(
 	if (!request.entry) {
 		throw new Error("Marketplace entry is required.")
 	}
-	return installMarketplaceEntryWithCli(request.entry)
+	return installMarketplaceEntryFromCatalog(request.entry)
 }
