+++ apps/vscode/src/core/controller/marketplace/marketplace-helpers.ts
@@ -6,6 +6,8 @@ import { basename, dirname, extname, isAbsolute, join, relative, resolve } from 
 import {
 	disablePluginMcpServersInSettings,
 	discoverPluginModulePaths,
+	installMcpServer,
+	installPlugin,
 	readGlobalSettings,
 	resolvePluginConfigSearchPaths,
 	setDisabledPlugin,
@@ -39,7 +41,6 @@ const MARKETPLACE_CATALOG_URL = "https://cline.github.io/marketplace/catalog.jso
 const OFFICIAL_PLUGINS_REPO = "https://github.com/cline/plugins.git"
 const INSTALL_COMMAND_TIMEOUT_MS = 120_000
 const MAX_OUTPUT_CHARS = 12_000
-const LOCAL_CLI_ENTRYPOINT_ENV = "CLINE_MARKETPLACE_CLI_PATH"
 const SECRET_PATTERN =
 	/(api[_ -]?key|access[_ -]?token|refresh[_ -]?token|auth(?:orization)?[_ -]?token|token|secret|password|authorization|credential)/i
 const SECRET_KEY_VALUE_PATTERN =
@@ -313,60 +314,77 @@ async function runCommand(command: string, args: string[]): Promise<SpawnResult>
 	})
 }
 
-function localCliRunner(): { command: string; args: string[] } | undefined {
-	const overridePath = process.env[LOCAL_CLI_ENTRYPOINT_ENV]?.trim()
-	const devWorkspacePath = process.env.DEV_WORKSPACE_FOLDER?.trim()
-	const candidatePath =
-		overridePath ||
-		(devWorkspacePath ? join(devWorkspacePath, "apps", "cli", "src", "index.ts") : undefined) ||
-		findLocalCliEntrypointFromKnownDirectories()
-	if (!candidatePath || !existsSync(candidatePath)) return undefined
-	return {
-		command: "bun",
-		args: ["--conditions=development", candidatePath],
+function parseMcpInstallArgs(args: string[]): {
+	name: string
+	headers: string[]
+	targetArgs: string[]
+	transport?: string
+} {
+	const [name, ...rest] = args
+	if (!name) throw new Error("Marketplace MCP install args must start with a server name.")
+	const headers: string[] = []
+	const targetArgs: string[] = []
+	let transport: string | undefined
+	for (let index = 0; index < rest.length; index++) {
+		const arg = rest[index]
+		if (arg === "--transport") {
+			const value = rest[index + 1]
+			if (!value) throw new Error("--transport requires a value")
+			transport = value
+			index++
+			continue
+		}
+		if (arg?.startsWith("--transport=")) {
+			transport = arg.slice("--transport=".length)
+			continue
+		}
+		if (arg === "--header") {
+			const value = rest[index + 1]
+			if (!value) throw new Error("--header requires a value")
+			headers.push(value)
+			index++
+			continue
+		}
+		if (arg?.startsWith("--header=")) {
+			headers.push(arg.slice("--header=".length))
+			continue
+		}
+		targetArgs.push(arg)
 	}
+	return { name, headers, targetArgs, transport }
 }
 
-function findLocalCliEntrypointFromKnownDirectories(): string | undefined {
-	const startDirectories = [process.cwd(), typeof __dirname === "string" ? __dirname : undefined].filter(
-		(directory): directory is string => Boolean(directory),
+function installMcpMarketplaceEntry(entry: MarketplaceEntry, args: string[]): MarketplaceInstallResult {
+	const parsed = parseMcpInstallArgs(args)
+	const result = installMcpServer(parsed)
+	return MarketplaceInstallResult.create({
+		id: entry.id,
+		type: entry.type,
+		status: "installed",
+		message: `Installed ${entry.name || entry.id}.`,
+		output: result.warnings.join("\n") || undefined,
+	})
+}
+
+async function installPluginMarketplaceEntry(entry: MarketplaceEntry, args: string[]): Promise<MarketplaceInstallResult> {
+	const [source] = args
+	if (!source) throw new Error("Marketplace plugin install args must start with a plugin source.")
+	const result = await installPlugin({ source })
+	const warnings = result.mcpSyncFailures.map(
+		(failure) => `Failed to sync plugin MCP servers for ${failure.pluginName ?? failure.pluginPath}: ${failure.message}`,
 	)
-	for (const startDirectory of startDirectories) {
-		const candidatePath = findLocalCliEntrypoint(startDirectory)
-		if (candidatePath) return candidatePath
-	}
-	return undefined
+	return MarketplaceInstallResult.create({
+		id: entry.id,
+		type: entry.type,
+		status: "installed",
+		message: `Installed ${entry.name || entry.id}.`,
+		output: [`Path: ${result.installPath}`, ...warnings].join("\n"),
+	})
 }
 
-function findLocalCliEntrypoint(startDirectory: string): string | undefined {
-	let current = resolve(startDirectory)
-	for (let depth = 0; depth < 8; depth++) {
-		const candidatePath = join(current, "apps", "cli", "src", "index.ts")
-		if (existsSync(candidatePath)) return candidatePath
-		const parent = dirname(current)
-		if (parent === current) break
-		current = parent
-	}
-	return undefined
-}
-
-export async function installMarketplaceEntryWithCli(entry: MarketplaceEntry): Promise<MarketplaceInstallResult> {
-	const args = getEntryArgs(entry)
-	if (args.length === 0) throw new Error("Marketplace install args are required.")
-	const localCli = entry.type === "mcp" || entry.type === "plugin" ? localCliRunner() : undefined
-	const command = localCli?.command ?? "npx"
-	const commandArgs = localCli
-		? [
-				...localCli.args,
-				...(entry.type === "mcp"
-					? ["mcp", "install", "--yes", "--json", ...args]
-					: ["plugin", "install", args[0] ?? "", "--json"]),
-			]
-		: entry.type === "skill"
-			? ["-y", "skills@latest", "add", ...args, "-g", "-a", "cline", "-y"]
-			: entry.type === "mcp"
-				? ["-y", "cline", "mcp", "install", "--yes", "--json", ...args]
-				: ["-y", "cline", "plugin", "install", args[0] ?? "", "--json"]
+async function installSkillMarketplaceEntry(entry: MarketplaceEntry, args: string[]): Promise<MarketplaceInstallResult> {
+	const command = "npx"
+	const commandArgs = ["-y", "skills@latest", "add", ...args, "-g", "-a", "cline", "-y"]
 	const displayCommand = formatCommand(command, commandArgs)
 	let result: SpawnResult
 	try {
@@ -393,6 +411,14 @@ export async function installMarketplaceEntryWithCli(entry: MarketplaceEntry): P
 		message: `Installed ${entry.name || entry.id}.`,
 		output,
 	})
+}
+
+export async function installMarketplaceEntryFromCatalog(entry: MarketplaceEntry): Promise<MarketplaceInstallResult> {
+	const args = getEntryArgs(entry)
+	if (args.length === 0) throw new Error("Marketplace install args are required.")
+	if (entry.type === "mcp") return installMcpMarketplaceEntry(entry, args)
+	if (entry.type === "plugin") return installPluginMarketplaceEntry(entry, args)
+	return installSkillMarketplaceEntry(entry, args)
 }
 
 function readPackageName(packageJsonPath: string): string | undefined {
